chore: create queryBuilderThunks to store thunks for query builder
